python-mcp/requirements.txt
mcp>=1.0.0
docker>=7.1.0
python-dotenv>=1.0.1
openai>=1.14.0
pydantic>=2.9.0
httpx>=0.27.0
